move process_message to its own crate
