+++ svm-message-processor/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "solana-svm-message-processor"
+description = "Solana VM message processor."
+documentation = "https://docs.rs/solana-svm-message-processor"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+solana-measure = { workspace = true }
+solana-program-runtime = { workspace = true }
+solana-sdk = { workspace = true }
+solana-svm-transaction = { workspace = true }
+solana-timings = { workspace = true }
+
+[dev-dependencies]
+bincode = { workspace = true }
+libsecp256k1 = { workspace = true }
+rand = { workspace = true }
+serde = { workspace = true }
+serde_derive = { workspace = true }
+solana-compute-budget = { workspace = true }
+solana-feature-set = { workspace = true }
+solana-sdk = { workspace = true, features = ["dev-context-only-utils"] }
+solana-type-overrides = { workspace = true }
+
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[lints]
+workspace = true
